Move lints related to must_use to their own module
